+++ lab01/bit_ops.h
@@ -1,3 +1,3 @@
-unsigned get_bit(unsigned x, unsigned n);
-void set_bit(unsigned * x, unsigned n, unsigned v);
-void flip_bit(unsigned * x, unsigned n);
+uint16_t  get_bit(uint16_t  x, uint16_t  n);
+void set_bit(uint16_t  * x, uint16_t  n, uint16_t  v);
+void flip_bit(uint16_t  * x, uint16_t  n);
